Skinner added

+++ Software/Werwolf/Werwolf/Printing/Job.cs
@@ -49,6 +49,13 @@ namespace Werwolf.Printing
         public bool MachBilder { get; set; }
         public bool Rotieren { get; set; }
         public bool KonsolenAnzeigen { get; set; }
+        /// <summary>
+        /// Soll die Job-Datei nach Ausführung gelöscht werden?
+        /// </summary>
+        public bool CleanJob { get; set; }
+        public int MaxCPU { get; set; } = 1;
+
+        public string UniversePath { get; set; }
 
         public Job(Universe Universe, string Pfad)
             : this()
@@ -70,7 +77,7 @@ namespace Werwolf.Printing
             bool FixedFont, bool TrennlinieVorne, bool TrennlinieHinten,
             bool Rotieren,
             float MaximaleGrose, bool MachBilder,
-            bool KonsolenAnzeigen)
+            bool KonsolenAnzeigen, bool CleanJob)
         {
             this.Deck = Deck;
             this.HintergrundFarbe = HintergrundFarbe;
@@ -85,6 +92,7 @@ namespace Werwolf.Printing
             this.MachBilder = MachBilder;
             this.Rotieren = Rotieren;
             this.KonsolenAnzeigen = KonsolenAnzeigen;
+            this.CleanJob = CleanJob;
 
             this.Init(Deck.Universe);
             this.Name = this.Schreibname = Deck.Name + "-Job";
@@ -106,6 +114,9 @@ namespace Werwolf.Printing
             this.MachBilder = Loader.XmlReader.GetBoolean("MachBilder");
             this.Rotieren = Loader.XmlReader.GetBoolean("Rotieren");
             this.KonsolenAnzeigen = Loader.XmlReader.GetBoolean("KonsolenAnzeigen");
+            this.CleanJob = Loader.XmlReader.GetBoolean("CleanJob");
+
+            this.UniversePath = Loader.XmlReader.GetString("Universe_Path");//For Skinner
         }
         protected override void WriteIntern(System.Xml.XmlWriter XmlWriter)
         {
@@ -124,6 +135,9 @@ namespace Werwolf.Printing
             XmlWriter.WriteBoolean("MachBilder", MachBilder);
             XmlWriter.WriteBoolean("Rotieren", Rotieren);
             XmlWriter.WriteBoolean("KonsolenAnzeigen", KonsolenAnzeigen);
+            XmlWriter.WriteBoolean("CleanJob", CleanJob);
+
+            XmlWriter.WriteAttribute("Universe_Path", Universe.Pfad);//For Skinner
         }
 
         public override void AdaptToCard(Karte Karte)
@@ -146,8 +160,91 @@ namespace Werwolf.Printing
                 kartenSize = kartenSize.permut();
             return WolfSinglePaper.GetNumberOfCards(kartenSize).Inhalt();
         }
+        //public void DistributedPrint(string JobPath, ProgressBar ProgressBar)
+        //    => DistributedPrint(JobPath, new JobTickerProgressBar(ProgressBar));
+        //{
+        //    int solvedJobs = 0;
+        //    int numberOfJobs;
+        //    if (MachBilder)
+        //        numberOfJobs = Deck.UniqueCount();
+        //    else
+        //    {
+        //        int numberProSheet = GetNumberProSheet();
+        //        int numberOfCards = Deck.TotalCount();
+        //        numberOfJobs = (int)Math.Ceiling(numberOfCards * 1f / numberProSheet);
+        //        if (MyMode == Job.RuckBildMode.Einzeln)
+        //            numberOfJobs *= 2;
+        //    }
 
-        public void DistributedPrint(string TargetPath, ProgressBar progressBar1)
+        //    if (progressBar1 != null)
+        //        progressBar1.Invoke((MethodInvoker)delegate
+        //        {
+        //            progressBar1.Value = 0;
+        //            progressBar1.Maximum = numberOfJobs;
+        //        });
+
+        //    Queue<int> jobs = new Queue<int>();
+        //    for (int i = 0; i < numberOfJobs; i++)
+        //        jobs.Enqueue(i);
+
+        //    int maxWorkers = FastMath.Min(Settings.MaximumNumberOfCores, MaxCPU, numberOfJobs, Environment.ProcessorCount);
+        //    Process[] workers = new Process[maxWorkers];
+
+        //    ProcessStartInfo psi = new ProcessStartInfo();
+        //    psi.FileName = Path.Combine(Directory.GetCurrentDirectory(), "WolfSlave.exe");
+        //    psi.Arguments = '"' + Universe.Pfad + "\" \"" + JobPath + "\"";
+
+        //    while (solvedJobs < numberOfJobs)
+        //    {
+        //        for (int i = 0; i < maxWorkers; i++)
+        //            if (workers[i] == null && jobs.Count > 0)
+        //                workers[i] = Process.Start(psi.FileName, psi.Arguments + " " + jobs.Dequeue());
+        //            else if (workers[i] != null && workers[i].HasExited)
+        //            {
+        //                solvedJobs++;
+        //                workers[i].Close();
+        //                workers[i] = null;
+        //                if (progressBar1 != null)
+        //                    progressBar1.Invoke((MethodInvoker)delegate { progressBar1.PerformStep(); });
+        //            }
+        //        Thread.Sleep(500);
+        //    }
+        //    if (!MachBilder)
+        //    {
+        //        string[] files = new string[numberOfJobs];
+        //        for (int i = 0; i < numberOfJobs; i++)
+        //            files[i] = Path.Combine(Path.GetDirectoryName(JobPath), Schreibname + "." + i + ".pdf");
+
+        //        bool pdfCreated = false;
+        //        while (!pdfCreated)
+        //        {
+        //            try
+        //            {
+        //                if (MyMode == RuckBildMode.Einzeln)
+        //                    PDFHelper.ConcatSplitDoppelseitig(Path.Combine(Path.GetDirectoryName(JobPath), Schreibname), (long)(MaximaleGrose * (1 << 20)), files);
+        //                else
+        //                    PDFHelper.ConcatSplit(Path.Combine(Path.GetDirectoryName(JobPath), Schreibname), (long)(MaximaleGrose * (1 << 20)), files);
+        //                pdfCreated = true;
+        //            }
+        //            catch (Exception e)
+        //            {
+        //                DialogResult dr = MessageBox.Show("Datei " + Path.Combine(Path.GetDirectoryName(JobPath), Schreibname + ".pdf")
+        //                      + " kann nicht erstellt werden. Bitte schließen Sie das Dokument und führen Sie diesen Vorgang nochmal aus."
+        //                      + "\r\nFehlernachricht:\r\n"
+        //                      + e.Message,
+        //                    "Dokument muss geschlossen werden",
+        //                    MessageBoxButtons.AbortRetryIgnore);
+        //                if (dr != DialogResult.Retry)
+        //                    pdfCreated = true;
+        //            }
+        //        }
+        //        foreach (var item in files)
+        //            File.Delete(item);
+        //    }
+        //    if (CleanJob)
+        //        File.Delete(JobPath);
+        //}
+        public void DistributedPrint(string JobPath, IJobTicker Ticker)
         {
             int solvedJobs = 0;
             int numberOfJobs;
@@ -158,32 +255,18 @@ namespace Werwolf.Printing
                 int numberProSheet = GetNumberProSheet();
                 int numberOfCards = Deck.TotalCount();
                 numberOfJobs = (int)Math.Ceiling(numberOfCards * 1f / numberProSheet);
-                if (MyMode == Printing.Job.RuckBildMode.Einzeln)
+                if (MyMode == Job.RuckBildMode.Einzeln)
                     numberOfJobs *= 2;
             }
-
-            if (progressBar1 != null)
-                progressBar1.Invoke((MethodInvoker)delegate
-                {
-                    progressBar1.Value = 0;
-                    progressBar1.Maximum = numberOfJobs;
-                });
-
-            string JobPath = Path.Combine(Path.GetDirectoryName(TargetPath), Schreibname + ".job.xml");
-            using (System.Xml.XmlWriter writer = System.Xml.XmlWriter.Create(JobPath))
-            {
-                writer.WriteStartDocument();
-                Write(writer);
-                writer.WriteEndDocument();
-                writer.Close();
-            }
+            Ticker.Reset(numberOfJobs);
 
             Queue<int> jobs = new Queue<int>();
             for (int i = 0; i < numberOfJobs; i++)
                 jobs.Enqueue(i);
 
-            int maxWorkers = FastMath.Min(Settings.MaximumNumberOfCores, numberOfJobs, Environment.ProcessorCount);
+            int maxWorkers = FastMath.Min(Settings.MaximumNumberOfCores, MaxCPU, numberOfJobs, Environment.ProcessorCount);
             Process[] workers = new Process[maxWorkers];
+            int[] jobIDs = new int[maxWorkers];
 
             ProcessStartInfo psi = new ProcessStartInfo();
             psi.FileName = Path.Combine(Directory.GetCurrentDirectory(), "WolfSlave.exe");
@@ -193,14 +276,16 @@ namespace Werwolf.Printing
             {
                 for (int i = 0; i < maxWorkers; i++)
                     if (workers[i] == null && jobs.Count > 0)
-                        workers[i] = Process.Start(psi.FileName, psi.Arguments + " " + jobs.Dequeue());
+                    {
+                        jobIDs[i] = jobs.Dequeue();
+                        workers[i] = Process.Start(psi.FileName, psi.Arguments + " " + jobIDs[i]);
+                    }
                     else if (workers[i] != null && workers[i].HasExited)
                     {
                         solvedJobs++;
+                        Ticker.Exited(jobIDs[i], workers[i].ExitCode);
                         workers[i].Close();
                         workers[i] = null;
-                        if (progressBar1 != null)
-                            progressBar1.Invoke((MethodInvoker)delegate { progressBar1.PerformStep(); });
                     }
                 Thread.Sleep(500);
             }
@@ -236,7 +321,20 @@ namespace Werwolf.Printing
                 foreach (var item in files)
                     File.Delete(item);
             }
-            File.Delete(JobPath);
+            if (CleanJob)
+                File.Delete(JobPath);
+        }
+        public string Save(string TargetPath)
+        {
+            string JobPath = Path.Combine(Path.GetDirectoryName(TargetPath), Schreibname + ".job.xml");
+            using (System.Xml.XmlWriter writer = System.Xml.XmlWriter.Create(JobPath))
+            {
+                writer.WriteStartDocument();
+                Write(writer);
+                writer.WriteEndDocument();
+                writer.Close();
+            }
+            return JobPath;
         }
     }
 }
